elasticache: Do not add primary host to the list of readers
In our current AWS provider fork we use a new attribute called
`read_endpoint_addresses`. This attribute contains a list of all the
hosts on the cluster. This means that primary (read/write) and replicas
(read) hosts are in the list. As such, clients would perform read
operations on the primary end point whilst others are writing to it.

This is not optimal so lets fix the attribute to only contain the
replica nodes.